fix bug

+++ controllers/user.js
@@ -683,7 +683,9 @@ exports.getUsers = async (req, res) => {
 
 exports.deleteUser = async (req, res) => {
 	try {
-		console.log(req.body);
+		if (!req.body.userId) {
+			return res.status(400).json(errormessage('UserId is required'));
+		}
 		const find = await User.findByIdAndDelete(req.body.userId);
 		res.status(200).json(successmessage('Deleted Successfully', find));
 	} catch (err) {
